Edited Create and Added Two DELETE API routes
Edited create_menuitems to include calories and allergens. The allergens would be passed in by their id number, I could change it in the future if we need.

Also added two different DELETE API routes delete_menuitem and delete_menuitem_by_id. First one takes in a menu item name and deletes it. If there are multiple with the same name it deletes the new one. The second one takes in the menu_item_id and deletes it. *Neither work if the item has already been ordered yet. Do not how we want to delete an item that has already been ordered before on associated with a transaction, causes an error due to key references.